feat: added SubscanService
