+++ tests/services/TxQueryService.test.ts
@@ -1,21 +1,17 @@
-import { FirebaseServiceMock } from './FirebaseServiceMock';
-import { IFirebaseService } from './../../src/services/FirebaseService';
-import { ApiFactory, IApiFactory } from '../../src/client/ApiFactory';
 import { TxQueryService } from '../../src/services/TxQueryService';
-import { AstarApiMock } from './AstarApiMock';
+import { ISubscanService } from './../../src/services/SubscanService';
+import { SubscanServiceMock } from './SubscanServiceMock';
 
 describe('fetch the transaction details', () => {
-    let apiFactory: IApiFactory;
-    let firebaseService: IFirebaseService;
+    let subscanMock: ISubscanService;
 
     beforeEach(() => {
-        apiFactory = new ApiFactory();
-        apiFactory.getApiInstance = jest.fn().mockReturnValue(new AstarApiMock());
-        firebaseService = new FirebaseServiceMock();
+        subscanMock = new SubscanServiceMock();
     });
 
     it('return the transfer details', async () => {
-        const service = new TxQueryService(firebaseService);
+        const service = new TxQueryService(subscanMock);
+        // const hash = '0x66b432feb410848602bfbdf5be34eff4124c820f3ab2492c59768dd4c58';
         const hash = '0x66b432feb410848602bfbdf5be34eff4124c820f3ab2492c59768dd4c58e92a0';
         const result = await service.fetchTransferDetails('shiden', hash);
         const { amount, from, to, symbol, timestamp, isSuccess } = result;
